scripts/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
feedparser==6.0.11
schedule==1.2.2
pytz==2024.2
